Make rk4_step importable
