+++ examples/euler/acoustic_pulse.py
@@ -40,6 +40,7 @@ from grudge.models.euler import (
     EulerOperator,
     InviscidWallBC
 )
+from grudge.shortcuts import rk4_step
 
 from meshmode.mesh import BTAG_ALL
 
@@ -49,14 +50,6 @@ import grudge.op as op
 
 import logging
 logger = logging.getLogger(__name__)
-
-
-def rk4_step(y, t, h, f):
-    k1 = f(t, y)
-    k2 = f(t+h/2, y + h/2*k1)
-    k3 = f(t+h/2, y + h/2*k2)
-    k4 = f(t+h, y + h*k3)
-    return y + h/6*(k1 + 2*k2 + 2*k3 + k4)
 
 
 def gaussian_profile(
