+++ examples/euler/vortex.py
@@ -37,6 +37,7 @@ from grudge.models.euler import (
     primitive_to_conservative_vars,
     EulerOperator
 )
+from grudge.shortcuts import rk4_step
 
 from pytools.obj_array import make_obj_array
 
@@ -44,14 +45,6 @@ import grudge.op as op
 
 import logging
 logger = logging.getLogger(__name__)
-
-
-def rk4_step(y, t, h, f):
-    k1 = f(t, y)
-    k2 = f(t+h/2, y + h/2*k1)
-    k3 = f(t+h/2, y + h/2*k2)
-    k4 = f(t+h, y + h*k3)
-    return y + h/6*(k1 + 2*k2 + 2*k3 + k4)
 
 
 def vortex_initial_condition(x_vec, t=0):
