+++ examples/wave/wave-op-var-velocity.py
@@ -40,7 +40,7 @@ from meshmode.mesh import BTAG_ALL, BTAG_NONE  # noqa
 
 from grudge.discretization import DiscretizationCollection
 from grudge.dof_desc import DISCR_TAG_BASE, DISCR_TAG_QUAD, DOFDesc
-from grudge.shortcuts import make_visualizer
+from grudge.shortcuts import make_visualizer, rk4_step
 
 import grudge.op as op
 
@@ -120,14 +120,6 @@ def wave_operator(dcoll, c, w):
     )
 
 # }}}
-
-
-def rk4_step(y, t, h, f):
-    k1 = f(t, y)
-    k2 = f(t+h/2, y + h/2*k1)
-    k3 = f(t+h/2, y + h/2*k2)
-    k4 = f(t+h, y + h*k3)
-    return y + h/6*(k1 + 2*k2 + 2*k3 + k4)
 
 
 def estimate_rk4_timestep(actx, dcoll, c):
